implement draft version of rendering
git-svn-id: https://osmand.googlecode.com/svn/trunk@481 e29c36b1-1cfa-d876-8d93-3434fc2bb7b8

+++ DataExtractionOSM/src/net/osmand/osm/MapUtils.java
@@ -198,11 +198,11 @@ public class MapUtils {
 		return x / getPowZoom(zoom) * 360.0 - 180.0;
 	}
 	
-	private static float getPowZoom(float zoom){
+	private static double getPowZoom(float zoom){
 		if(zoom - Math.ceil(zoom) < 0.05f){
 			return 1 << ((int)zoom); 
 		} else {
-			return (float) Math.pow(2, zoom);
+			return Math.pow(2, zoom);
 		}
 	}
